requirements.txt
streamlit~=1.41.1
requests~=2.32.3
numpy~=2.1.3
scikit-learn~=1.5.2
matplotlib~=3.9.2
openpyxl~=3.1.5
python-dotenv~=1.0.1